created a file for namespace
