+++ include/PointsInfo.h
@@ -8,7 +8,7 @@
 /// \date April 30th, 2022
 ///
 
-#include "ezc3d.h"
+#include "ezc3dNamespace.h"
 
 ///
 /// \brief 3D rotation data
