+++ include/math/Matrix.h
@@ -8,7 +8,7 @@
 /// \date March 25th, 2020
 ///
 
-#include "ezc3d.h"
+#include "ezc3dNamespace.h"
 
 ///
 /// \brief Matrix of unknown dimension
