+++ src/Header.cpp
@@ -8,6 +8,7 @@
 ///
 
 #include "Header.h"
+#include "ezc3d.h"
 #include "Parameters.h"
 #include "DataStartInfo.h"
